Use correct type for invoiceItemSubscription
If you look at the stripe API for invoice items https://stripe.com/docs/api#invoiceitem_object,
the subscription value of an invoice item is the subscriptionId as a string.  But the type of
`invoiceItemSubscription` field is `Subscription` so we are trying to parse a string like "sub_xxxx"
into a Subscription.  To fix tis, the type of `invoiceItemSubscription` must change to `SubscriptionId`
and `SubscriptionId` needs a `FromJSON` instance.