Fix dependencies and versions in Nix expressions
The version numbers in all of the default.nix files were lagging behind
the version numbers stated in the cabal files. This mismatch was causing
some confusing behaviour when I tried to build these packages locally.

I refreshed the versions and dependencies by running cabal2nix in each
of the sub-projects.